fix: change PlayerData.Level type from float to int

+++ Runtime/Player/PlayerData.cs
@@ -4,6 +4,6 @@ namespace AchEngine.Player
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public float Level { get; set; }
+        public int Level { get; set; }
     }
 }
